+++ Watchdogd/Include/Dls.h
@@ -73,11 +73,11 @@
 
  extern void Run_dls ( void );                                                          /* Dans The_dls.c */
  extern int EA_inrange( int num );
- extern double EA_ech( int num );
+ extern float EA_ech( int num );
  extern int A( int num );
  extern char *TRdetail( int num );
  extern void SE( int num, int etat );
- extern void SEA( int num, double val_avant_ech );
+ extern void SEA( int num, float val_avant_ech );
  extern void SEA_range( int num, int range );
  extern void SEA_ech( int num, float val_ech );
  extern void Envoyer_commande_dls ( int num );
